+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.2.0
 
+* New Keyboard::leds_mut function for getting underlying leds object.
+
 Breaking changes:
 * Update to generic_array 0.14, that is exposed in matrix. The update
   should be transparent.